pyproject.toml
[project]
name = "mcp-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.18.0",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
]
